rename timeline feature to time
