+++ knowde/_feature/__init__.py
@@ -3,4 +3,4 @@ from .location import *  # noqa: F403
 from .proposition import *  # noqa: F403
 from .reference import *  # noqa: F403
 from .sentence import *  # noqa: F403
-from .timeline import *  # noqa: F403
+from .time import *  # noqa: F403
