+++ knowde/_feature/time/domain/errors.py
