+++ knowde/cli.py
@@ -5,7 +5,7 @@ import click
 
 from knowde._feature import prop_cli, ref_cli, tl_cli
 from knowde._feature.location.interface import loc_cli
-from knowde.feature import deduct_cli, def_cli
+from knowde.feature import deduct_cli, def_cli, event_cli
 from knowde.feature.person.interface import person_cli
 from knowde.reference.interface import refdef_cli
 
@@ -32,3 +32,4 @@ cli.add_command(deduct_cli)
 cli.add_command(tl_cli)
 cli.add_command(loc_cli)
 cli.add_command(person_cli)
+cli.add_command(event_cli)
