+++ knowde/feature/event/__init__.py
@@ -5,4 +5,4 @@
 * Eventと同レベルの依存度なので同レベルのフィーチャーにすべし
 * よってこのフィーチャーではPersonへ依存させない
 """
-from .interface import ev_router  # noqa: F401
+from .interface import ev_router, event_cli  # noqa: F401
