+++ knowde/feature/event/domain/event.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 
 from knowde._feature._shared.domain.domain import Entity
 from knowde._feature.location.domain import Location  # noqa: TCH001
-from knowde._feature.timeline.domain.domain import Time  # noqa: TCH001
+from knowde._feature.time.domain.domain import Time  # noqa: TCH001
 
 
 class Event(Entity, frozen=True):
