+++ knowde/feature/space_time/__init__.py
@@ -1,0 +1,1 @@
+"""時間と場所に依存した機能."""
